Database connection and Yusufs player profile page
Added Yusufs playerProfile page and CSS alterations.

Established the connection to the database. (IMPORTANT) If you want to connect to the database made on your local web server, change the details in the config.ini file in configurations. Only change the username and password as your database name should be the same.