emergency changes made prior to sem 1 presentation.
Changes to make profile page dynamic. 

emergency changes that means maple execution is not queued but rather is simply updated immediately. (this will need to be rectified as will cause lag for the user, however, was required for demo purposes). This is because Utas block changes to /etc/at.allow

Grant please review before merging.