Bookmarking
bookmaking uses js to create / amend cookie with player id.
A php function has been created in content-manager.php to retrieve their info from that database. An array is returned containing player details. From here it is trivial to use this to fill in the bookmarked player information into the appropriate page.